+++ src/main/java/skatn/remindmeback/common/security/controller/dto/LogoutRequest.java
@@ -1,0 +1,4 @@
+package skatn.remindmeback.common.security.controller.dto;
+
+public record LogoutRequest(String refreshToken) {
+}
