+++ src/main/java/skatn/remindmeback/common/security/controller/dto/ReissueTokenRequest.java
@@ -1,0 +1,4 @@
+package skatn.remindmeback.common.security.controller.dto;
+
+public record ReissueTokenRequest(String refreshToken) {
+}
