+++ src/main/java/skatn/remindmeback/common/security/entity/RefreshToken.java
@@ -1,0 +1,28 @@
+package skatn.remindmeback.common.security.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.*;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class RefreshToken {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String token;
+    private String parentToken;
+    private LocalDateTime expiration;
+    private String tokenGroup;
+    private Long memberId;
+
+}
