+++ src/main/java/skatn/remindmeback/common/security/filter/RestAuthenticationFilter.java
@@ -1,0 +1,54 @@
+package skatn.remindmeback.common.security.filter;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.http.HttpMethod;
+import org.springframework.security.authentication.AuthenticationServiceException;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.authentication.AbstractAuthenticationProcessingFilter;
+import org.springframework.security.web.context.DelegatingSecurityContextRepository;
+import org.springframework.security.web.context.RequestAttributeSecurityContextRepository;
+import org.springframework.security.web.context.SecurityContextRepository;
+import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
+import org.springframework.util.StringUtils;
+import skatn.remindmeback.common.security.dto.AccountDto;
+import skatn.remindmeback.common.security.token.RestAuthenticationToken;
+
+import java.io.IOException;
+
+public class RestAuthenticationFilter extends AbstractAuthenticationProcessingFilter {
+
+    private final ObjectMapper objectMapper = new ObjectMapper();
+
+    public RestAuthenticationFilter() {
+        super(new AntPathRequestMatcher("/api/login", "POST"));
+    }
+
+    public SecurityContextRepository getSecurityContextRepository(HttpSecurity http) {
+        SecurityContextRepository securityContextRepository = http.getSharedObject(SecurityContextRepository.class);
+        if(securityContextRepository == null) {
+            securityContextRepository = new DelegatingSecurityContextRepository(new RequestAttributeSecurityContextRepository());
+        }
+
+        return securityContextRepository;
+    }
+
+    @Override
+    public Authentication attemptAuthentication(HttpServletRequest request, HttpServletResponse response) throws AuthenticationException, IOException {
+        if(!HttpMethod.POST.name().equals(request.getMethod())) {
+            throw new IllegalArgumentException("Authentication method not supported");
+        }
+
+        AccountDto accountDto = objectMapper.readValue(request.getReader(), AccountDto.class);
+
+        if (!StringUtils.hasText(accountDto.username()) || !StringUtils.hasText(accountDto.password())) {
+            throw new AuthenticationServiceException("Username or Password not provided");
+        }
+        RestAuthenticationToken token = new RestAuthenticationToken(accountDto.username(), accountDto.password());
+
+        return getAuthenticationManager().authenticate(token);
+    }
+}
