+++ src/main/java/skatn/remindmeback/common/security/handler/OAuth2AuthenticationFailureHandler.java
@@ -1,0 +1,24 @@
+package skatn.remindmeback.common.security.handler;
+
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.authentication.AuthenticationFailureHandler;
+import org.springframework.stereotype.Component;
+import skatn.remindmeback.common.security.dto.OAuth2Properties;
+
+import java.io.IOException;
+
+@Component
+@RequiredArgsConstructor
+public class OAuth2AuthenticationFailureHandler implements AuthenticationFailureHandler {
+
+    private final OAuth2Properties oAuth2Properties;
+
+    @Override
+    public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response, AuthenticationException exception) throws IOException {
+        response.sendRedirect(oAuth2Properties.failureRedirectUrl());
+    }
+}
