+++ src/main/java/skatn/remindmeback/common/security/jwt/JwtProperties.java
@@ -1,0 +1,11 @@
+package skatn.remindmeback.common.security.jwt;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties("auth.jwt")
+public record JwtProperties(
+        String secret,
+        long accessTokenExpireTime,
+        long refreshTokenExpireTime
+) {
+}
