+++ src/main/java/skatn/remindmeback/common/security/jwt/TokenDto.java
@@ -1,0 +1,4 @@
+package skatn.remindmeback.common.security.jwt;
+
+public record TokenDto(String accessToken, String refreshToken) {
+}
