+++ src/main/java/skatn/remindmeback/common/security/provider/RestAuthenticationProvider.java
@@ -1,0 +1,38 @@
+package skatn.remindmeback.common.security.provider;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.authentication.AuthenticationProvider;
+import org.springframework.security.authentication.BadCredentialsException;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Component;
+import skatn.remindmeback.common.security.dto.AccountContext;
+import skatn.remindmeback.common.security.token.RestAuthenticationToken;
+
+@Component
+@RequiredArgsConstructor
+public class RestAuthenticationProvider implements AuthenticationProvider {
+
+    private final UserDetailsService userDetailsService;
+    private final PasswordEncoder passwordEncoder;
+
+    @Override
+    public Authentication authenticate(Authentication authentication) throws AuthenticationException {
+        String email = authentication.getName();
+        String password = (String) authentication.getCredentials();
+        AccountContext accountContext = (AccountContext) userDetailsService.loadUserByUsername(email);
+
+        if(!passwordEncoder.matches(password, accountContext.getPassword())) {
+            throw new BadCredentialsException("Invalid password");
+        }
+
+        return new RestAuthenticationToken(accountContext.getAccountDto(), null, accountContext.getAuthorities());
+    }
+
+    @Override
+    public boolean supports(Class<?> authentication) {
+        return authentication.isAssignableFrom(RestAuthenticationToken.class);
+    }
+}
