--- src/main/java/skatn/remindmeback/security/controller/dto/LogoutRequest.java
@@ -1,4 +1,0 @@
-package skatn.remindmeback.security.controller.dto;
-
-public record LogoutRequest(String refreshToken) {
-}
