--- src/main/java/skatn/remindmeback/security/controller/dto/ReissueTokenRequest.java
@@ -1,4 +1,0 @@
-package skatn.remindmeback.security.controller.dto;
-
-public record ReissueTokenRequest(String refreshToken) {
-}
