--- src/main/java/skatn/remindmeback/security/dsl/RestApiDsl.java
@@ -1,64 +1,0 @@
-package skatn.remindmeback.security.dsl;
-
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.annotation.web.HttpSecurityBuilder;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractAuthenticationFilterConfigurer;
-import org.springframework.security.web.authentication.AuthenticationFailureHandler;
-import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
-import org.springframework.security.web.authentication.RememberMeServices;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.security.web.authentication.session.SessionAuthenticationStrategy;
-import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
-import org.springframework.security.web.util.matcher.RequestMatcher;
-import skatn.remindmeback.security.filter.RestAuthenticationFilter;
-
-public class RestApiDsl <H extends HttpSecurityBuilder<H>> extends AbstractAuthenticationFilterConfigurer<H, RestApiDsl<H>, RestAuthenticationFilter> {
-
-    private AuthenticationSuccessHandler successHandler;
-    private AuthenticationFailureHandler failureHandler;
-
-    public RestApiDsl(){
-        super(new RestAuthenticationFilter(),null);
-    }
-
-    @Override
-    public void configure(H http) {
-
-        AuthenticationManager authenticationManager = http.getSharedObject(AuthenticationManager.class);
-        getAuthenticationFilter().setAuthenticationManager(authenticationManager);
-        getAuthenticationFilter().setAuthenticationSuccessHandler(successHandler);
-        getAuthenticationFilter().setAuthenticationFailureHandler(failureHandler);
-        getAuthenticationFilter().setSecurityContextRepository(getAuthenticationFilter().getSecurityContextRepository((HttpSecurity) http));
-
-        SessionAuthenticationStrategy sessionAuthenticationStrategy = http.getSharedObject(SessionAuthenticationStrategy.class);
-        if (sessionAuthenticationStrategy != null) {
-            getAuthenticationFilter().setSessionAuthenticationStrategy(sessionAuthenticationStrategy);
-        }
-        RememberMeServices rememberMeServices = http.getSharedObject(RememberMeServices.class);
-        if (rememberMeServices != null) {
-            getAuthenticationFilter().setRememberMeServices(rememberMeServices);
-        }
-        http.setSharedObject(RestAuthenticationFilter.class,getAuthenticationFilter());
-        http.addFilterBefore(getAuthenticationFilter(), UsernamePasswordAuthenticationFilter.class);
-    }
-
-    public RestApiDsl<H> restSuccessHandler(AuthenticationSuccessHandler successHandler) {
-        this.successHandler = successHandler;
-        return this;
-    }
-
-    public RestApiDsl<H> restFailureHandler(AuthenticationFailureHandler authenticationFailureHandler) {
-        this.failureHandler = authenticationFailureHandler;
-        return this;
-    }
-
-    public RestApiDsl<H> loginPage(String loginPage) {
-        return super.loginPage(loginPage);
-    }
-
-    @Override
-    protected RequestMatcher createLoginProcessingUrlMatcher(String loginProcessingUrl) {
-        return new AntPathRequestMatcher(loginProcessingUrl, "POST");
-    }
-}
