--- src/main/java/skatn/remindmeback/security/entity/RefreshToken.java
@@ -1,28 +1,0 @@
-package skatn.remindmeback.security.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.*;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class RefreshToken {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String token;
-    private String parentToken;
-    private LocalDateTime expiration;
-    private String tokenGroup;
-    private Long memberId;
-
-}
