--- src/main/java/skatn/remindmeback/security/handler/RestAuthenticationFailureHandler.java
@@ -1,25 +1,0 @@
-package skatn.remindmeback.security.handler;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.authentication.AuthenticationFailureHandler;
-import org.springframework.stereotype.Component;
-import skatn.remindmeback.common.controlleradvice.dto.ErrorResponse;
-import skatn.remindmeback.common.exception.ErrorCode;
-
-import java.io.IOException;
-
-@Component
-public class RestAuthenticationFailureHandler implements AuthenticationFailureHandler {
-
-    private final ObjectMapper objectMapper = new ObjectMapper();
-
-    @Override
-    public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response, AuthenticationException exception) throws IOException {
-        response.setStatus(ErrorCode.INVALID_USERNAME_OR_PASSWORD.getStatus());
-        response.setContentType("application/json;charset=UTF-8");
-        objectMapper.writeValue(response.getWriter(), new ErrorResponse(ErrorCode.INVALID_USERNAME_OR_PASSWORD));
-    }
-}
