--- src/main/java/skatn/remindmeback/security/jwt/JwtProperties.java
@@ -1,11 +1,0 @@
-package skatn.remindmeback.security.jwt;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties("auth.jwt")
-public record JwtProperties(
-        String secret,
-        long accessTokenExpireTime,
-        long refreshTokenExpireTime
-) {
-}
