--- src/main/java/skatn/remindmeback/security/jwt/TokenDto.java
@@ -1,4 +1,0 @@
-package skatn.remindmeback.security.jwt;
-
-public record TokenDto(String accessToken, String refreshToken) {
-}
