--- src/main/java/skatn/remindmeback/security/token/RestAuthenticationToken.java
@@ -1,36 +1,0 @@
-package skatn.remindmeback.security.token;
-
-import org.springframework.security.authentication.AbstractAuthenticationToken;
-import org.springframework.security.core.GrantedAuthority;
-
-import java.util.Collection;
-
-public class RestAuthenticationToken extends AbstractAuthenticationToken {
-
-    private final Object principal;
-    private final Object credentials;
-
-    public RestAuthenticationToken(Object principal, Object credentials, Collection<? extends GrantedAuthority> authorities) {
-        super(authorities);
-        this.principal = principal;
-        this.credentials = credentials;
-        setAuthenticated(true);
-    }
-
-    public RestAuthenticationToken(Object principal, Object credentials) {
-        super(null);
-        this.principal = principal;
-        this.credentials = credentials;
-        setAuthenticated(false);
-    }
-
-    @Override
-    public Object getCredentials() {
-        return credentials;
-    }
-
-    @Override
-    public Object getPrincipal() {
-        return principal;
-    }
-}
